Let <Esc> transfer focus from Board to ICS Input
<Esc> in the input field of the ICS Interaction window transfers focus
to the board, but there was no way to transfer focus back without actually
typing something in the input field. <Esc> now does that. Unlike typing
printables, it does not close the chat pane, though. In addition, <Esc>
when the chat pane is open now also transfers focus to the board, rather
than closing the chat pane. This makes quick transparent switching between
board and Chat / ICS Interaction possible. It is no longer possible to
simply hide the chat pane, though. But this was usually done for typing
a command, and swicthing to the board with <Esc> and typing the command
there has the same effect.