Recognize castling double-moves from engine
An engine could already send double-moves of the Alien protocol,
where the same piece moved twice, which were then glued into a single
step, with the intermediate square as trampled piece. Now when the
second leg is whith a different piece, it keeps the first leg as overall
move, and tramples the second mover. This translates castlings sent
as two-piece moves to the kludge format of trampling the own 'Rook'.