Take heed of mnemonic indicator when clipping menu texts
An underscore in the menu text should not be counted when deciding where
to clip the text to make the menu bar fit the window width. If clipping
would occur immediately after an underscore, just clip off the first
character to get the mnemonic back in view. (This makes _n from the
Engine menu.)