Man and Info Page Fix
Terminal.app doesn’t accept command line arguments. Rather then use OSX
feature of using .command scripts to launch the man and info pages
(which is problematic with gatekeeper),  it is much better to open them
directly with applescript.   Had to move infoProc to xboard.c to do
this though.