Allow Lion double-moves in opening book
Double moves can now be encoded in a Polyglot es promoType = 2, and
a code for two King steps in the to-square. The moves are retrieved
in the Alien Edition protocol format (i.e. comma-separated LAN).
HandleMachineMove had to be adapted to handle the two legs coming as one,
but otherwise the book code can just send the retrieved line directly
to the engine.
 For Edit Book the format has to be converted to TSA-like for display,
to make the move parser that will read it on saving understand the move
as a single one.