Also allow Princess SVG piece to be diversify
The image for the Grasshopper has a copy in the Chu-promoted series.
The unpromoted image (not needed in Chu) is now preferentially taken
from a Crown image rather than Princess, allowing distinction between
the promoted and unpromoted piece in kanji themes for Wa Shogi.
Any piece between Grasshopper and chu-promoted Marhall can now in principle
use a backup.