Make Claws glyph available in non-Chu variants
The variant-dependent replacement glyphs for promotable pieces in Chu
are now assigned to pieces that reuse the replaced glyphs to represent
promoted pieces, rather than to unused promoted pieces. This means that
in Chu they now must be replaced by swapping, which avoids occurrence
of duplicats both in Chu and other variants. And thus frees piece types
that now can be used for new glyphs. As a result the Nightrider remains
available in Chu, and the Claw glyph in other variants. Other freed
pieces are for now used for the promotion Golds of variant shogi.