Correctly remember checkboxes on Continue Later (WB)
The WinBoard tournament dialog could be left in an inconsistent state
on Continue Later, when the user untickedthe checkboxes for auto-stepping
and two games per line, after on an earlier attempt their ticking had
already resulted in settng the line/position index to a negative value.
Negative values left in the numeric entriesare now reset to 0 when the
chackboxes are unticked.