Prevent printing in non-existing Chat dialog (XB)
The Xaw build of XBoard crashed in -ics -zp mode when the opponent was
kibitzing. According to a stack trace OutputChatMessage() tried to change
the color of a non-existent button. Although it was not clear how printing
there could be attempted if no chats were defined, the routine now is made
a no-op when the Chat dialog does not exist, by an immediate return.