Start implementing EPD test suites
An option -epd is added that can be used in combination with matchMode
to let XBoard do a test suite on a file of EPDs with a "bm" qualifier.
It would record the time to solution. Reporting still has to be implemented;
as yet unsolved problem is that currently black-to-move EPDs would
be presented to the second engine.