Try to load bitmaps for all pieces (WB)
WinBoard only contains build-in bitmaps for the first 22 piece types,
and also only associated filenames with those that can be used to
provide external images. Now a new naming convention for piece images
is introduced, like piece22_49s.bmp for a 49x49 bitmap for solid piece 22.
All pieces that were not loaded according to the old convention now
try to find images with this naming convention if a -pieceImageDirectory
is defined.