Add duplicat of Lion (and Flying Dragon)
Two more piece types are added: a duplicat of the Lion in the unpromoted
series, and a Flying Dragon as its promoted counterpart. This Lion does
not suffer from the Lion-trading restriction of Chu Shogi, and can be used
as a general piece. Its pictogram is somewhat smaller than that of the
Chu Shogi Lion.