Correct backup pieces for addition of minor Lion
The added piece type was not put in the list of backup pieces, so that
the pieces beyond it got out of phase, and no backup for the Prince
would be found, leading to an error popup about missing default pieces.
The error message on this has now been improved to mention the piece
casing the trouble.