Let FENs handle Betza initial rights in castlingless variants
When a variant has no castling, or is redefined to have no castling
through a 'piece' command for the King, the FEN castling field of a
written FEN will now contain the file ID of all pieces with initial
('i') moves in their Betza definition. On reading the virginity of
pieces not thus mentioned in the castling field will get their
virginity revoked.