Add support for Multi-PV Margin
An engine option with a name containing multi, PV and margin is now
recognized, and will lead to printing of a multi-PV header (during
analysis in the Engine Output pane) even in absence of a MultiPV option.
Right-clicking the part of that header between 'fewer' and 'more'
allows tuning of the option value through horizontal mouse moving.
During this tuning the value is displayed in the message field of the
main window. When there is no MultiPV option to specify the number of
PVs, the margin value will be displayed in the header instead of the
number of PVs.