Remove added pango modules to coincide with macports package
Macports got rid of the +builtin_modules variant and made that the default port.  (I assume macports, maybe it was a gtk change)  This is great for installation simplicity, but messes up my build script which was set to use the former base pango install. Hopefully they stick with this.