Use mouse wheel for selecting piece in Edit Position mode (XB)
Rather than calling ForwardEvent() and BackwardEvent() directly from the
mouse-wheel handler, we now call a backend routine Wheel(), which gets
the pointer coordinates passed to it (which we must remember during
mouse move events). This then calls the Forward/Backward event, except
in EditPosition mode, where it changes the piece in the square below the
pointer.