Fix bug #43792 (no highlights after rejection of premove)
If the sending of a premove fails to add a move to the game, because
XBoard itself judges the move illegal (which could only be determined
after arrival of the opponent move, not when it was entered), XBoard
now re-applies the move highlights for that opponent move.
 This does not solve the problem yet with legality testing off,
when it later turns out the ICS rejects the move.