Add 'divide by 60' checkbox in Time Control dialog XB
The TC could not be set to fractional minutes (base time) or seconds
(increment) in the GTK version, as the spin controls there do not allow
you to type such numbers. An added checkbox 'Divide entered times by 60'
allows you to enter base time in seconds, and increment in 1/60 sec.
The dialog always pops upwith this checkbox unticked.