Fix black border around saved diagrams (WB)
The 'Save as Diagram' function of WinBoard produced bitmaps with a black
border, because the buffer bitmap used for drawing the board (which was
flushed to file to create the diagram) had a size larger than the board,
including somemargins of the main window. Now we copy it to a board-sized
bitmap first, and flush that.