Silence warning
Conversion to use of strncpy had led to the length of a (char*) being
tested, and atypically we got away with that so far because the copied
string was only 2 characters long. Now we use length limit 4, which also
renders the use of strncpy instead of strcpy useless, but at least
does notproduce a warning.