Fix parent dialog of Error Popup
If the parent dialog of the Error Popup is wrong, the OK button in it
will not manage to pop it down. Migrating the Tourney Options to their
own DialogClass had changed testing for a parent other than the board
window from TransientDlg to MasterDlg, but Toerney Options was not the
only such dialog throwing up errors. (E.g. New Variant coulddo it too.)
So we now test if MasterDlg or TransientDlg is up, and use those as
parent if they are.