Make OK and Cancel buttons appear in top-level dialogs GTK
The OK and Cancel buttons are created in GTK as part of the
dialog_with_buttons call, but to get top-level windows another call
had to be used, which does not automatically provide them.
We now explicitly add the buttons as requested by the EndMark Option,
paying attention to the SAME_ROW requests. (The code for that is currently
a bit messy, as one suspects that the cumbersome testing on Option.type
was done exactly to make the SAME_ROW request of EndMarks ignored.)
The GTK "clicked" signal is translated to a "response" in a kludgy way,
through a PopDownProxy routine.