Cleanup Edit Tags/Book/EngineList a bit
The Tags dialog is also used for two other functions, and this hack
causes some problems. This patch ameliorates these a bit. The Edit
Engine List now displays its own title (instead of "Tags"), and popping
it up pops down which other function was in use first. (Unfortunately
this can result in data loss for uncommitted changes.) The name of the
"save changes" button has been changed to "commit changes", to suggest
less strongly that a change in the Engine List would be immediately
saved to file (which in reality requires Save Settings Now as well).