Improve triple-leg-move animation
When both locust squares (kill and kill2) are set, animation now takes
place in three stages. As kill2 is not retreived when stepping through
the game history, this works only for moves just entered or played by
the engine. Replayed moves are still animated in 2 stages, though, which
does look pretty good on 2-out-1-in moves.