Deprecate -defaultPathEGTB option
The value of the -defaultPathEGTB option is now automatically cleared
at startup, after appending it as "nalimov:" format to -egtFormats.
We also no longer preconfigure a (dummy) value for -defaultPathEGTB,
but instead one for -egtFormats, (which will show up in the EGT path
field of the Common Engine Settings dialog now -defaultPathEGTB is
empty) to serve as an example for the syntax of -egtFormats.