Logout from ICS after fatal error
Exiting of XBoard after a fatal error such as crashing of the engine
could be delayed by popping up an exit message, which then had to be
confirmed first. This would leave XBoard in zippy mode open for new
challenges, which it would even accept, and lose by forfeit. Now
we send a 'logout' command to the ICS before popping up the exit
message.