Implement help clicks
A button3 click on the text labels of text-edit, spin, checkbox and
button controls, as well as menu items, now pops up an alert box
with the name of the clicked item. This as a first step to making an
interactive help that would use that name to look up information on the
item in the texi file, and display that.