Implement rough help popup
A routine is added that searches a given text header in the man file,
and then displays the following text until the next header (after
skipping additional headers that come before any text) in a message
box. The text uses a simple fill and wrap to line length > 80, and
ignores all control characters, except empty lines. Both .B and .SS
directives are recognized as headers.