Fix popdown of menus on help click
Because menus by default also react to button3 clicks, using those
clicks for help popup needs to suppress the default action. But this
action was also reponsible for the popdown of the menu, so that it
stayed up, covering the help dialog. Now we do allow the default action,
but flag to the generic menu handler that it should not execute the
menu function.