Make dialog labels and comboboxes also accept help clicks
Combobaoxes and bare labels had their own code for creating the label,
nd were not yet made sensitive to help clicks. (Which is a bit cumbersome,
as GTK does not allow attaching callbacks to label widgets, so that you
have to wrap them first into something that does.) The bare labels could
furthermore already have a callback specified (used for the clocks);
in this case, and the case of multi-line labels, no help handler is
attached.