Also try to get help for engine options
Help clicks in the Engine Settings dialogs will try to scan the man page
of the corresponding engine for info about the option. This is a bit
tricky, as it must assume the name of the engine is the same as that
of the engine command and the manual for it is in section 6. In addition
it must look in both /usr/share/man and /usr/local/share/man, as it cannot
know whether the engine was installed from source or repo. Finally it
must hope the man page isn't gzipped.