Suppress empty label at top of Edit Tags dialog
There is a Label Option above the text memo of the Edit Tags dialog
that can display a message, but it is only used in cmail mode, and
is passed as a NULL pointer otherwise. (On which GenericPopup then
crashed when it wanted to attach a help click.) Now we simply suppress
display of that widget when there is no message.