Implement XBetza iso modifier
An 'i' modifier on a continuation leg will force it to be of equal length
(measured in steps) as the latest sliding leg (i.e. of maximum length != 1)
before it. This is useful for defining rifle captures of sliding pieces.
On initial legs it still means the move is for virgin pieces only.