Also provide help on adapter options
Adapters often add their own options to thoso of the engine in the
Engine Settings dialog. Polyglot, for instance, has many options,
their names all starting with "Polyglot ...". We adopt this as a
convention, so that XBoard can compare the adapter name (the name
of the program actually run, if that is different from the name
of the engine binary) with the start of the option name for which
help is requested. If there is a match it will conclude we are dealing
with an adapter option, and will consult the adapter manual (for an
option without the adapter-name prefix) instead of the engine manual.