Streamline XBoard board drawing
XBoard was drawing in a very inefficient way. To improve efficiency
alteration of highlights will not immediately be exposed, but be
marked as damage so that a subsequent DrawPosition call will show them.
Squares touched by a to-be-erased arrow will be treated likewise.
The drawing now has three passes: repairing damage (which erases
highlights and arrow), emptying the from-square, and drawing new
pieces and highlights.