Repair flashing of moved piece (XB)
The flashing was not working (at least in GTK), because the redraw of
the square needed even processing to show something. So the draw/erase
loop that does the flashing needs to call DoEvents() to make things
visible. This however could lead to recursive processing of LeftClick,
as the press of a click-click move already enters the move and starts
the flashing, so that the release usually comes before the press
processing has finished. Which would cause it to be interpreted as
a second move identical to the first. A static flag inside LeftClick
now makes it ignore clicks during flashing, which is just what we had
to do with the release anyway. Flashing should be suppressed during
sweep-selection under-promotion.