Make startup announcement self-disappearing
The startup message is now configurable through an option -startupMessage,
and the master config file sets this to announce the help clicks.
The message can contain multiple lines. The display of lines that are a
sub-string of the option -messageSuppress is suppressed, however.
Using a help-click sets this latter option to the help-click announcement,
so that people who have used help clicks won't see this message anymoe.
Displaying the startup message is now only considered on board draws
that ordered a full repaint.