Prevent FICS bell character fro printing in ICS Console XB
FICS sends bell characters (ascii 7) at the start of many messages.
We now strip a leading bell off the message before appending it to
the text memo of the ICS Console, and ignore the message altogether
if it was just bell + linefeed (to prevent each move of an observed
game adding an empty line).