Use the official GTK font selector
A gtk_font_button is now put in the Fonts dialog to allow selection
of a general font. The individual window fonts got a button marked *
to assign this font to them. To implement the selector button some
kludgy code had to be added in the generic dialog constructor, to
recognize these Button Options by name, and use the special GTK
widget to implement them, or assign an unusual callback.