Adjust window height after clock-font change
A routine LockBoardSize are added, and called to protect the board's
Graph widget from absorbing the height change pf clock and message field.
Enlarging these now forces the top-level window to expand, and after the
change we try to reduce this window size to tightly fit the board to
squeeze out empty area after shrinking.